Move PeerMessageReceiverState into PeerMessageReceiver
